# showcase/symfony/patched/public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Job Board API",
        "description": "API for job listings aggregated from multiple sources. Generated from PHP typed arrays and array shapes.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/jobs/{id}": {
            "get": {
                "summary": "Action to get a single job by ID.",
                "operationId": "getJob",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Job ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "company_name": {
                                            "type": "string"
                                        },
                                        "company_logo": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "location": {
                                            "type": "string"
                                        },
                                        "remote": {
                                            "type": "boolean"
                                        },
                                        "job_type": {
                                            "type": "string"
                                        },
                                        "salary": {
                                            "type": "object",
                                            "properties": {
                                                "min": {
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "max": {
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "formatted": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "formatted"
                                            ]
                                        },
                                        "url": {
                                            "type": "string"
                                        },
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "source": {
                                            "type": "string"
                                        },
                                        "posted_at": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "fetched_at": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "title",
                                        "company_name",
                                        "location",
                                        "remote",
                                        "job_type",
                                        "salary",
                                        "url",
                                        "tags",
                                        "source",
                                        "description"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/jobs": {
            "get": {
                "summary": "Action to list jobs with filtering and pagination.",
                "operationId": "listJobs",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Search query"
                    },
                    {
                        "name": "remote",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        },
                        "description": "Filter remote jobs"
                    },
                    {
                        "name": "job_type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "full-time",
                                "part-time",
                                "contract",
                                "internship"
                            ]
                        },
                        "description": "Job type filter"
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Location filter"
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Source filter"
                    },
                    {
                        "name": "min_salary",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Minimum salary"
                    },
                    {
                        "name": "max_salary",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Maximum salary"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "Page number"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100
                        },
                        "description": "Items per page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "company_name": {
                                                        "type": "string"
                                                    },
                                                    "company_logo": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "location": {
                                                        "type": "string"
                                                    },
                                                    "remote": {
                                                        "type": "boolean"
                                                    },
                                                    "job_type": {
                                                        "type": "string"
                                                    },
                                                    "salary": {
                                                        "type": "object",
                                                        "properties": {
                                                            "min": {
                                                                "type": "integer",
                                                                "nullable": true
                                                            },
                                                            "max": {
                                                                "type": "integer",
                                                                "nullable": true
                                                            },
                                                            "currency": {
                                                                "type": "string",
                                                                "nullable": true
                                                            },
                                                            "formatted": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "required": [
                                                            "formatted"
                                                        ]
                                                    },
                                                    "url": {
                                                        "type": "string"
                                                    },
                                                    "tags": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "source": {
                                                        "type": "string"
                                                    },
                                                    "posted_at": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                },
                                                "required": [
                                                    "id",
                                                    "title",
                                                    "company_name",
                                                    "location",
                                                    "remote",
                                                    "job_type",
                                                    "salary",
                                                    "url",
                                                    "tags",
                                                    "source"
                                                ]
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "required": [
                                                "current_page",
                                                "per_page",
                                                "total",
                                                "last_page"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/jobs/stats": {
            "get": {
                "summary": "Action to get job statistics.",
                "operationId": "getJobStats",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "total_jobs": {
                                            "type": "integer"
                                        },
                                        "by_source": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "integer"
                                            }
                                        },
                                        "by_type": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "integer"
                                            }
                                        },
                                        "remote_jobs": {
                                            "type": "integer"
                                        },
                                        "with_salary": {
                                            "type": "integer"
                                        },
                                        "last_fetched_at": {
                                            "type": "string",
                                            "nullable": true
                                        }
                                    },
                                    "required": [
                                        "total_jobs",
                                        "by_source",
                                        "by_type",
                                        "remote_jobs",
                                        "with_salary"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": []
    }
}